requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.45.2
anyio==4.8.0
arxiv==2.2.0
async-timeout==4.0.3
attrs==24.3.0
Bottleneck==1.5.0
cachetools==5.5.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
dataclasses-json==0.6.7
deprecation==2.1.0
distro==1.9.0
docstring_parser==0.16
exceptiongroup==1.2.2
fastapi==0.115.0
feedparser==6.0.11
frozenlist==1.5.0
google-api-core==2.24.2
google-api-python-client==2.175.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.93.1
google-cloud-bigquery==3.33.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gotrue==2.11.1
greenlet==3.1.1
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
httpx-sse==0.4.0
hyperframe==6.0.1
idna==3.10
iniconfig==2.1.0
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.2
langchain-anthropic==0.3.7
langchain-community==0.3.1
langchain-core==0.3.60
langchain-experimental==0.3.4
langchain-google-community==2.0.7
langchain-google-vertexai==2.0.10
langchain-openai==0.2.1
langchain-text-splitters==0.3.0
langgraph==0.4.8
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.51
langsmith==0.1.147
marshmallow==3.25.1
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
numexpr==2.10.2
numpy==1.26.4
openai==1.59.7
orjson==3.10.14
ormsgpack==1.10.0
packaging==24.2
pillow==11.2.1
pipdeptree==2.26.1
pluggy==1.6.0
postgrest==0.17.2
propcache==0.2.1
proto-plus==1.26.1
protobuf==5.29.4
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.7.1
pydantic_core==2.23.4
Pygments==2.19.2
PyMuPDF==1.25.5
pymupdf4llm==0.0.24
pyparsing==3.2.3
PyPDF2==3.0.1
pytesseract==0.3.13
pytest==8.4.1
pytest-asyncio==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
realtime==2.1.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
setuptools==78.1.0
sgmllib3k==1.0.0
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.38.6
storage3==0.8.2
StrEnum==0.4.15
supabase==2.8.1
supabase_auth==2.11.3
supafunc==0.6.2
tenacity==8.5.0
tiktoken==0.8.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
uritemplate==4.2.0
urllib3==2.3.0
uvicorn==0.34.0
validators==0.35.0
websockets==13.1
wheel==0.45.1
xxhash==3.5.0
yarl==1.18.3
